+++ R/spct.utils.r
@@ -237,19 +237,22 @@ thin_wl.default <- function(x, ...) {
 #' @describeIn thin_wl
 #'
 #' @param max.wl.step numeric. Largest allowed wavelength difference between
-#'    adjacent spectral values in nanometres (nm).
+#'   adjacent spectral values in nanometres (nm).
 #' @param max.slope.delta numeric in 0 to 1. Largest allowed change in relative
-#'    slope of the spectral quantity per nm betweem adjacent pairs of values.
+#'   slope of the spectral quantity per nm between adjacent pairs of values.
+#' @param span integer A peak (or valley) is defined as an element in a sequence
+#'   which is greater (or smaller) than all other elements within a window of
+#'   width span centred at that element. Use NULL for the global peak.
 #' @param col.names character. Name of the column of \code{x} containing the
 #'   spectral data to check against \code{max.slope.delta}. Currently only one
 #'   column supported.
 #'
 #' @note The value of \code{max.slope.delta} is expressed as relative change in
 #'   the slope of spectral variable per nanometre. This means that values
-#'   between 0.0005 and 0.005 tend to work reasonably well. The best value
-#'   will depend on the wavelength step of the input and noise in data. A
-#'   moderate smoothing before thinning can sometimes help in the case of
-#'   noisy data.
+#'   between 0.0005 and 0.005 tend to work reasonably well. The best value will
+#'   depend on the wavelength step of the input and noise in data. A moderate
+#'   smoothing before thinning can sometimes help in the case of noisy data.
+#'
 #'   The amount of thinning is almost always less than the value of criteria
 #'   passed as argument as it is based on existing wavelength values. For
 #'   example if we start with a spectrum with a uniform wavelength step of 1 nm,
@@ -259,11 +262,15 @@ thin_wl.default <- function(x, ...) {
 #'   the algorithm attempts not to discard information, contrary to smoothing or
 #'   interpolation.
 #'
+#'   Local peaks and valleys are always preserved, using by default a span of 21
+#'   to search for them. See \code{\link{find_peaks}}.
+#'
 #' @export
 #'
 thin_wl.generic_spct <- function(x,
                                  max.wl.step = 10.0,
                                  max.slope.delta = 0.001,
+                                 span = 21,
                                  col.names,
                                  ...) {
   # compute stopping criterion
@@ -281,8 +288,8 @@ thin_wl.generic_spct <- function(x,
                                    col.names = col.names,
                                    na.rm = FALSE)
   # collect peaks and valleys to ensure that they are not removed
-  peaks <- find_peaks(x.norm[[col.names]], span = 21, strict = FALSE)
-  valleys <- find_peaks(-x.norm[[col.names]], span = 21, strict = FALSE)
+  peaks <- find_peaks(x.norm[[col.names]], span = span, strict = FALSE)
+  valleys <- find_peaks(-x.norm[[col.names]], span = span, strict = FALSE)
   extremes <- peaks | valleys
 
   # iterative loop
@@ -324,18 +331,21 @@ thin_wl.generic_spct <- function(x,
 thin_wl.source_spct <- function(x,
                                 max.wl.step = 10.0,
                                 max.slope.delta = 0.001,
+                                span = 21,
                                 unit.out = getOption("photobiology.radiation.unit", default = "energy"),
                                 ...) {
   if (unit.out == "energy") {
     thin_wl.generic_spct(x = q2e(x, action = "replace"),
                          max.wl.step = max.wl.step,
                          max.slope.delta = max.slope.delta,
+                         span = span,
                          col.names = "s.e.irrad",
                          ...)
   } else if (unit.out %in% c("photon", "quantum")) {
     thin_wl.generic_spct(x = e2q(x, action = "replace"),
                          max.wl.step = max.wl.step,
                          max.slope.delta = max.slope.delta,
+                         span = span,
                          col.names = "s.q.irrad",
                          ...)
   } else {
@@ -350,18 +360,21 @@ thin_wl.source_spct <- function(x,
 thin_wl.response_spct <- function(x,
                                   max.wl.step = 10.0,
                                   max.slope.delta = 0.001,
+                                  span = 21,
                                   unit.out = getOption("photobiology.radiation.unit", default = "energy"),
                                   ...) {
   if (unit.out == "energy") {
     thin_wl.generic_spct(x = q2e(x, action = "replace"),
                          max.wl.step = max.wl.step,
                          max.slope.delta = max.slope.delta,
+                         span = span,
                          col.names = "s.e.response",
                          ...)
   } else if (unit.out %in% c("photon", "quantum")) {
     thin_wl.generic_spct(x = e2q(x, action = "replace"),
                          max.wl.step = max.wl.step,
                          max.slope.delta = max.slope.delta,
+                         span = span,
                          col.names = "s.q.response",
                          ...)
   } else {
@@ -378,22 +391,32 @@ thin_wl.response_spct <- function(x,
 thin_wl.filter_spct <- function(x,
                                 max.wl.step = 10.0,
                                 max.slope.delta = 0.001,
+                                span = 21,
                                 qty.out = getOption("photobiology.filter.qty",
                                                     default = "transmittance"),
                                 ...) {
   if (qty.out == "transmittance") {
-    thin_wl.generic_spct(x = A2T(x, action = "replace"),
+    thin_wl.generic_spct(x = any2T(x, action = "replace"),
                          max.wl.step = max.wl.step,
                          max.slope.delta = max.slope.delta,
+                         span = span,
                          col.names = "Tfr",
                          ...)
   } else if (qty.out == "absorbance") {
-    thin_wl.generic_spct(x = T2A(x, action = "replace"),
+    thin_wl.generic_spct(x = any2A(x, action = "replace"),
                          max.wl.step = max.wl.step,
                          max.slope.delta = max.slope.delta,
+                         span = span,
                          col.names = "A",
                          ...)
-  } else {
+  } else if (qty.out == "absorptance") {
+    thin_wl.generic_spct(x = any2Afr(x, action = "replace"),
+                         max.wl.step = max.wl.step,
+                         max.slope.delta = max.slope.delta,
+                         span = span,
+                         col.names = "Afr",
+                         ...)
+  }else {
     stop("'unit.out ", qty.out, " is unknown")
   }
 }
@@ -405,10 +428,12 @@ thin_wl.filter_spct <- function(x,
 thin_wl.reflector_spct <- function(x,
                                    max.wl.step = 10.0,
                                    max.slope.delta = 0.001,
+                                   span = 21,
                                    ...) {
   thin_wl.generic_spct(x = x,
                        max.wl.step = max.wl.step,
                        max.slope.delta = max.slope.delta,
+                       span = span,
                        col.names = "Rfr",
                        ...)
 }
@@ -420,6 +445,7 @@ thin_wl.reflector_spct <- function(x,
 thin_wl.solute_spct <- function(x,
                                 max.wl.step = 10.0,
                                 max.slope.delta = 0.001,
+                                span = 21,
                                 ...) {
   cols <- intersect(c("K.mole", "K.mass"), names(x))
   if (length(cols) == 1) {
@@ -430,6 +456,7 @@ thin_wl.solute_spct <- function(x,
   thin_wl.generic_spct(x = x,
                        max.wl.step = max.wl.step,
                        max.slope.delta = max.slope.delta,
+                       span = span,
                        col.names = col.name,
                        ...)
 }
@@ -471,11 +498,13 @@ thin_wl.calibration_spct <- thin_wl.default
 thin_wl.generic_mspct <- function(x,
                                   max.wl.step = 10.0,
                                   max.slope.delta = 0.001,
+                                  span = 21,
                                   ...) {
   msmsply(x,
           thin_wl,
           max.wl.step = max.wl.step,
           max.slope.delta = max.slope.delta,
+          span = span,
           ...)
 }