+++ data-raw/make.filter.data.r
@@ -19,20 +19,8 @@ setWhatMeasured(opaque.spct, "theoretical fully opaque object")
 library(photobiologyFilters)
 
 polyester.spct <- filters.mspct$McDermit_PET_Autostat_CT5_125um
-setTfrType(polyester.spct, "total")
-setWhatMeasured(polyester.spct, "clear polyester film")
-setFilterProperties(polyester.spct,
-                    Rfr.constant = 0.07, # guess
-                    thickness = 125e-6,
-                    attenuation.mode = "absorption")
 
 yellow_gel.spct <- filters.mspct$Rosco_Canary_Supergel_no312
-setTfrType(polyester.spct, "total")
-setWhatMeasured(yellow_gel.spct, "yellow theatrical 'gel', Rosco supergel no. 312, 'canary yellow'")
-setFilterProperties(yellow_gel.spct,
-                    Rfr.constant = 0.07, # guess
-                    thickness = 85e-6,
-                    attenuation.mode = "absorption")
 
 save(clear.spct, file = "./data/clear.spct.rda")
 save(opaque.spct, file = "./data/opaque.spct.rda")